+++ src/components/hamburger/Hamburger.scss
@@ -26,7 +26,7 @@
   background-size: 117px 188px;
   transition: all 0.4s cubic-bezier(0.42, 0, 0.34, 1.01);
   position: fixed;
-  z-index: 5;
+  z-index: 10;
   right: -100%;
   top: 0;
   height: 100%;
