+++ src/components/navbar/NavBar.css
@@ -2,7 +2,7 @@
   position: fixed;
   display: flex;
   width: 100%;
-  z-index: 100;
+  z-index: 5;
   top: 0%;
   height: 80px;
   justify-content: space-between;
